Get rid of AnySigner.h.
